Refactoring to add main folder under src
