+++ src/main/UI/Button.java
@@ -1,7 +1,6 @@
-package src.UI;
+package UI;
 
 import javax.swing.*;
-import java.awt.*;
 
 public class Button {
     JButton button;
