+++ src/main/UI/Icon.java
@@ -1,7 +1,6 @@
-package src.UI;
+package UI;
 
 import javax.swing.*;
-import java.awt.*;
 
 public class Icon {
     ImageIcon icon;
