+++ src/main/entities/Password.java
@@ -1,4 +1,4 @@
-package src.entities;
+package entities;
 
 public abstract class Password {
     String passwordValue;
