+++ src/main/entities/StrengthChecker.java
@@ -1,4 +1,4 @@
-package src.entities;
+package entities;
 
 public class StrengthChecker {
     Strength strength;
